+++ signing/src/main/java/tech/pegasys/web3signer/signing/secp256k1/azure/AzureKeyVaultSigner.java
@@ -12,9 +12,9 @@
  */
 package tech.pegasys.web3signer.signing.secp256k1.azure;
 
-import static tech.pegasys.web3signer.keystorage.azure.AzureKeyVault.createUsingClientSecretCredentials;
-
 import tech.pegasys.web3signer.keystorage.azure.AzureKeyVault;
+import tech.pegasys.web3signer.signing.config.AzureAuthenticationMode;
+import tech.pegasys.web3signer.signing.config.AzureKeyVaultFactory;
 import tech.pegasys.web3signer.signing.secp256k1.EthPublicKeyUtils;
 import tech.pegasys.web3signer.signing.secp256k1.Signature;
 import tech.pegasys.web3signer.signing.secp256k1.Signer;
@@ -40,13 +40,15 @@ public class AzureKeyVaultSigner implements Signer {
   private final AzureConfig config;
   private final ECPublicKey publicKey;
   private final SignatureAlgorithm signingAlgo;
+  private final AzureKeyVaultFactory azureKeyVaultFactory;
   private final boolean needsToHash; // Apply Hash.sha3(data) before signing
 
   AzureKeyVaultSigner(
       final AzureConfig config,
       final Bytes publicKey,
       final boolean needsToHash,
-      final boolean useDeprecatedSignatureAlgorithm) {
+      final boolean useDeprecatedSignatureAlgorithm,
+      final AzureKeyVaultFactory azureKeyVaultFactory) {
     this.config = config;
     this.publicKey = EthPublicKeyUtils.createPublicKey(publicKey);
     this.needsToHash = needsToHash;
@@ -54,6 +56,7 @@ public class AzureKeyVaultSigner implements Signer {
         useDeprecatedSignatureAlgorithm
             ? SignatureAlgorithm.fromString("ECDSA256")
             : SignatureAlgorithm.ES256K;
+    this.azureKeyVaultFactory = azureKeyVaultFactory;
   }
 
   @Override
@@ -61,11 +64,12 @@ public class AzureKeyVaultSigner implements Signer {
     final AzureKeyVault vault;
     try {
       vault =
-          createUsingClientSecretCredentials(
+          azureKeyVaultFactory.createAzureKeyVault(
               config.getClientId(),
               config.getClientSecret(),
+              config.getKeyVaultName(),
               config.getTenantId(),
-              config.getKeyVaultName());
+              AzureAuthenticationMode.CLIENT_SECRET);
     } catch (final Exception e) {
       LOG.error("Failed to connect to vault", e);
       throw new SignerInitializationException(INACCESSIBLE_KEY_ERROR, e);