+++ signing/src/testFixtures/java/tech/pegasys/web3signer/signing/config/DefaultAzureKeyVaultParameters.java
@@ -10,10 +10,7 @@
  * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
  * specific language governing permissions and limitations under the License.
  */
-package tech.pegasys.web3signer.dsl.utils;
-
-import tech.pegasys.web3signer.signing.config.AzureAuthenticationMode;
-import tech.pegasys.web3signer.signing.config.AzureKeyVaultParameters;
+package tech.pegasys.web3signer.signing.config;
 
 import java.util.Collections;
 import java.util.HashMap;